bs_type: mark all explicit templates instantiations as exported
this is strictly speaking always needed, but gcc & llvm can handle
such cases somehow and VS compiler complains about unresolved externals even if it
actully see function body.
The reason behind it is that compiler cannot unabiguousely figure
out, which translation unit actually contains instantiation of static member
function, so it looks for explicitly exported one.
To resolve that issue we must explicily instantite static function and mark it as
exported even if function definition is available to compiler.