node: fix deep node copy
Fix a bug that links in cloned node didn't have proper owner set. We
can't setup link's owner in `node` copy constructor, because nodes are
managed using shared pointers and no shared pointers to node is created
in copy constructor. As a consequence, we cannot call
`bs_shared_this()` and obtain a `shared_ptr` to `this` in node's
copy constructor.

To resolve an issue, add special postprocess function
`detail::adjust_cloned_node()` that is called by `type_descriptor`
right after an object is cloned. Implementation of this function checks
if an object is a `node` and if so, calls `node::propagate_owner()`
that properly sets an onwer in all contained links.