tree/link: `data_ex()` and `data_node_ex()` return error on fail
This is major link API change: introduce new pure virtual methods that
return error if corresponding operation failed for some reason. With new
API link can return context (error) describing WHY an operation failed.

Useful for symbolic and other links that acquire data on demand.

This commit is a naive change and there are some things to improve,
first of all -- return error instead of null pointer. Hence, if
`data_ex()` (for example) returns a valid pointer, then we can rely that it is
not null.