tree/error: switch to using tree error codes
Note to prev commit: there exists specific error code
`tree::Error::OKOK` that is NOT an error and instead is alternative
value of 'OK'.

This code is intended to be used by `fusion_iface` to signal that object
is fully loaded (populated and data obtained). If this value is returned
from `fusion_iface::populate()` or `fusion_iface::data_node()` then BOTH
`DataNode` and `Data` status are set to `OK` by `fusion_link`.