tree/link: add methods to set handle of pointee node to self
When a link to node A is inserted into parent node we need to set
handle of node A to the link being inserted (to be sure that only one
hard link exists to node A). Existing approach is to access node via
`link::data_node()` and set it's handle directly.

But calling `link::data_node()` might be inefficient for fusion and
other links with delayed structure loading.

This commit adds virtual protected API to link that allows to switch
pointee handle to self. This API is intended to be called by node when
postprocessing inserted links.