serial/tree: add very first implementation of Tree FS archive
Archive is developed to be the default for representing BS tree in
filesystem. It has the following design goals.

1. Nodes are mapped into filesystem directories.

2. Links are mapped into individual files inside parent node.

3. Objects are serialized into individual files inside dedicated
directory ('.objects' by default) in root tree folder.

4. Objects can be serialized independent from each other and in
parallel. But doing so comes at a cost: cross shared_ptrs between
objects are not tracked for uniqueness and can be dumped multiple times
in different object files. But that's intentional and acceptable.
Parallel serialization is not yet implemented.

5. Client code should be able to register save/load pairs fr different
formats of object representation. By default all objects are serialized
to portable binary archives.
Only prototype is implemented here.

Also current implementation has issues.

1. Custom links order inside node isn't preserved. This is not fixed for
future development purposes (but can be fixed easily).

2. Problem related to p. 1 and parallel loading - sometimes Cereal fails
when trying to deserialize unregistered shared_ptr. Fixing for this is
upcoming and will require more Cereal modification.