tree/link: add new `LinkDeleted` event that is called on link termination
Sometimes we want to execute something when link is really deleted and
not just erased from some node. Add specific event that will be triggered
when link actor is terminating.

Handler for this event will receive *NULL* link and will only be able to
obtain deleted link's ID. There is no way to access link itself because
it might be already destroyed.