python/kernel: export fn to control waiting for actors before shutdown
When registering Python callbacks as tree event handlers it may be
useful to NOT wait until all actors done in `caf::actor_system`
destructor.

What might happen:
1. Python callbacks are subscribed to events -> are put into CAF
   registry.
2. During exit kernel can be terminated soon, before collecting links &
   nodes variables visible in Python.
3. Kernel destructor calls `caf::actor_system` destructor which will
   by default wait until registry is empty.
4. Links and nodes terminate their group in destructor, but some of 'em
   still alive -> registry is not empty.
5. Steps 3 & 4 cause deadlock.

To avoid deadlock we can disable waiting for all actors done when
terminating an application. This seems like preferred mode for desktop
app termination as opposed to possible headless "server mode" when
waiting for all actors is a priority.