kernel/python: TEMP disable adapters cache cleanup on link delete
Sometimes it seems that instance of adapter is already deleted (!)
despite that at least one reference to it is stored inside BS cache (in
form of `pybind11::object`). When pe::object's destructor decrements
refcounter we get a segfault.

Maybe that's an issue with Python's cyclic GC that collects all objects
with refcounter < 2. Will investigate and fix this further.