tree: allow 'dirty' signal for DataNode
Because `node` is not related to `objbase` and `objnode` is a
composition, we can separately track changes in object itself and in
contained node.

This commit allows DataNode = OK signal to always fire regardless of
prev. state.

And node acks behavior forwards up DataNode = OK instead of Data = OK as
earlier.